backend/pyproject.toml
[project]
name = "joo-coin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "aiomysql>=0.3.2",
    "alembic>=1.16.5",
    "asyncmy>=0.2.10",
    "cryptography>=46.0.3",
    "fastapi>=0.121.3",
    "greenlet>=3.2.4",
    "openai>=2.7.1",
    "pydantic-settings>=2.11.0",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.1",
    "pyupbit>=0.2.34",
    "sqlalchemy>=2.0.44",
    "uv>=0.9.11",
    "uvicorn>=0.38.0",
    "apscheduler>=3.10.0",
    "ruff>=0.14.6",
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "pytest-cov>=5.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
pythonpath = ["."]
